refactor: move sortHand to deck
